Use surface-specific table cell constructors
